apps/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
anthropic==0.43.0
celery[redis]==5.4.0
redis[hiredis]==5.2.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
httpx==0.28.1
PyJWT==2.10.1
cryptography==44.0.0
python-dotenv==1.0.1
sse-starlette==2.2.1
